Fix stb_image issue #1861 by zero-initializing `palette`.
My guess is the cost of a 1024-element memset is probably less than checking the palette index of every pixel, even if the branches in the latter fix were always predicted correctly, because it prevents vectorization.

Also zero-initialize `tc16` for safety; I think it may be possible to use it uninitialized.